requirements.txt
openai
streamlit
pypdf
requests
beautifulsoup4
python-dotenv
